Remove multi-dim dynamic array filler-value check
This will not give syntax errors for multidimensional arrays with
dynamic arrays which does not need filler values. This allow user to
create fixed size (sealed) arrays with abstract objects but with inner
most dimension is dynamic array. Run time check will still run to make
sure dynamic automatic array fill by compiler is not allowed. User must
give valid objects at runtime to these arrays.

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/BIRGen.java
@@ -173,7 +173,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
 import javax.xml.XMLConstants;
 
 import static org.wso2.ballerinalang.compiler.desugar.AnnotationDesugar.ANNOTATION_DATA;
@@ -2055,7 +2054,7 @@ public class BIRGen extends BLangNodeVisitor {
         long size = -1L;
         if (listConstructorExpr.type.tag == TypeTags.ARRAY &&
                 ((BArrayType) listConstructorExpr.type).state != BArrayState.UNSEALED) {
-            size = (long) ((BArrayType) listConstructorExpr.type).size;
+            size = ((BArrayType) listConstructorExpr.type).size;
         } else if (listConstructorExpr.type.tag == TypeTags.TUPLE) {
             size = listConstructorExpr.exprs.size();
         }
